Use existing devopsclient
